# Optimize Version.match?/2 and Version.compare/2 (#9807)